Use float for storing temperature.

+++ include/model.h
@@ -6,7 +6,7 @@
 typedef struct {
   uint16_t pm25 = 0;
   uint16_t pm10 = 0;
-  uint16_t temp = 0;
+  float temp = 0;
   uint16_t humidity = 0;
   uint16_t pressure = 0;
 } JsonModel;
@@ -16,7 +16,7 @@ typedef struct {
   uint16_t pm25rel;
   uint16_t pm10;
   uint16_t pm10rel;
-  uint16_t temp;
+  float temp;
   uint16_t humidity;
   uint16_t pressure;
   float level;
